Decouple cli setup/implementation from main package
